LPS-192399 Move and rename base class to object-rest-test-util module
